+++ src/types/schema.d.ts
@@ -684,6 +684,8 @@ export interface components {
       containerId?: number | null;
       /** Location */
       location?: number | null;
+      /** Sub Location */
+      subLocation?: number | null;
       /** Details */
       details?: Record<string, never> | null;
       /** Comments */
@@ -703,6 +705,7 @@ export interface components {
       parent?: string | null;
       /** Type */
       type: string;
+      dataCollectionGroupId?: number | null;
     };
     /** ShipmentChildren */
     ShipmentChildren: {